+++ src/test/java/cz/upce/fei/inptp/project/entity/DistanceMatrixTest.java
@@ -1,4 +1,4 @@
-package cz.upce.fei.inptp.zz.entity;
+package cz.upce.fei.inptp.project.entity;
 
 import cz.upce.fei.inptp.project.entity.matrix.DistanceMatrix;
 import org.junit.Test;
