+++ src/test/java/cz/upce/fei/inptp/project/entity/DriverTest.java
@@ -1,8 +1,7 @@
-package cz.upce.fei.inptp.zz.entity;
+package cz.upce.fei.inptp.project.entity;
 
 import static org.junit.Assert.*;
 
-import cz.upce.fei.inptp.project.entity.Driver;
 import org.junit.Test;
 
 public class DriverTest {
