+++ src/test/java/cz/upce/fei/inptp/project/planner/PlanTest.java
@@ -1,8 +1,8 @@
-package cz.upce.fei.inptp.zz.planner;
+package cz.upce.fei.inptp.project.planner;
 
-import cz.upce.fei.inptp.zz.entity.Driver;
-import cz.upce.fei.inptp.zz.entity.Vehicle;
-import cz.upce.fei.inptp.zz.entity.VehicleType;
+import cz.upce.fei.inptp.project.entity.Driver;
+import cz.upce.fei.inptp.project.entity.vehicle.Vehicle;
+import cz.upce.fei.inptp.project.entity.vehicle.VehicleType;
 import java.util.List;
 import org.junit.Test;
 import static org.junit.Assert.*;
